+++ src/components/layout/header/button-list/ButtonList.tsx
@@ -5,7 +5,7 @@ import hair from '../../../../assets/icons/hair-care.svg';
 import heater from '../../../../assets/icons/heater.svg';
 import headphone from '../../../../assets/icons/headphone.svg';
 import light from '../../../../assets/icons/lighting.svg';
-import { INavItems } from '../header';
+import { INavItems } from '../Header';
 import { FastCatalog } from '../../../../shared/constants/text-fields';
 import {
   List,
