+++ src/pages/about/About.tsx
@@ -1,14 +1,7 @@
-import Footer from '../../components/layout/footer/Footer';
-import { Link } from 'react-router-dom';
-
 export const AboutPage = () => {
   return (
     <>
-      <h2>This About page</h2>
-      <div>
-        Go <Link to={'/'}>home</Link>
-      </div>
-      <Footer></Footer>
+      <div>this page About</div>
     </>
   );
 };
