+++ src/pages/not-found/Notfound.tsx
@@ -1,39 +1,9 @@
-import { useNavigate } from 'react-router-dom';
-import Footer from '../../components/layout/footer/Footer';
-
-import styles from './notFound.module.scss';
-import { Button } from '@mui/material';
+import { NotFound } from '../../components/layout/not-found/Notfound';
 
 export const NotFoundPage = () => {
-  const navigate = useNavigate();
-
-  const handleGoHome = () => {
-    navigate('/');
-  };
-
-  const handleGoBack = () => {
-    navigate(-1);
-  };
-
   return (
     <>
-      <main className={styles.main}>
-        <div className={styles.content}>
-          <h2 className={styles.pageHeader}>
-            Oops! This page
-            <br /> doesn’t exist.
-          </h2>
-          <div className={styles.buttonWrapper}>
-            <Button className={styles.button} onClick={handleGoHome}>
-              home
-            </Button>
-            <Button className={styles.button} onClick={handleGoBack}>
-              back
-            </Button>
-          </div>
-        </div>
-      </main>
-      <Footer></Footer>
+      <NotFound />
     </>
   );
 };
